requirements.txt
ultralytics>=8.0.0
segment-anything>=1.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
git+https://github.com/openai/CLIP.git#egg=clip
easyocr>=1.7.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.3.1
streamlit>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0
matplotlib>=3.7.1
scikit-image>=0.20.0
